Reduce _MSC_VER footprint
