+++ src/pc/psxsdk/libcard.c
@@ -1,8 +1,3 @@
-#ifdef _MSC_VER
-#include <SDL.h>
-#else
-#include <SDL2/SDL.h>
-#endif
 #include <common.h>
 #include <psxsdk/libcard.h>
 #include <log.h>
