+++ src/pc/sotn.c
@@ -2,11 +2,6 @@
 #include "dra.h"
 #include "servant.h"
 #include "sfx.h"
-#ifdef _MSC_VER
-#include <SDL.h>
-#else
-#include <SDL2/SDL.h>
-#endif
 #include <cJSON/cJSON.h>
 
 #include <stdio.h>
